Move the emojiPickerContent back to main index
